refactor: events permissions
